Removed warning

+++ qemu/target-ppc/translate_init.c
@@ -9068,8 +9068,8 @@ static int ppc_fixup_cpu(PowerPCCPU *cpu)
         || (env->insns_flags2 & ~PPC_TCG_INSNS2)) {
         fprintf(stderr, "Warning: Disabling some instructions which are not "
                 "emulated by TCG (0x%" PRIx64 ", 0x%" PRIx64 ")\n",
-                env->insns_flags & ~PPC_TCG_INSNS,
-                env->insns_flags2 & ~PPC_TCG_INSNS2);
+                env->insns_flags & (uint64_t)~PPC_TCG_INSNS,
+                env->insns_flags2 & (uint64_t)~PPC_TCG_INSNS2);
     }
     env->insns_flags &= PPC_TCG_INSNS;
     env->insns_flags2 &= PPC_TCG_INSNS2;
